ENG-2197: Type of "[limit]" and "[ttl]" bind variables should be INT32
Summary: Change "[limit]" and "[ttl]" bind variable type to INT32 to avoid incompatibility with Cassandra.

Test Plan: TestBindVariable.testBindingLimit and TestBindVariable.testBindingTTL

Reviewers: pritam.damania

Reviewed By: pritam.damania

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D3141

+++ java/yb-cql/src/test/java/org/yb/cql/BaseCQLTest.java
@@ -58,7 +58,7 @@ public class BaseCQLTest extends BaseMiniClusterTest {
   protected static final Logger LOG = LoggerFactory.getLogger(BaseCQLTest.class);
 
   // Integer.MAX_VALUE seconds is the maximum allowed TTL by Cassandra.
-  protected static final long MAX_TTL_SEC = Integer.MAX_VALUE;
+  protected static final int MAX_TTL_SEC = Integer.MAX_VALUE;
 
   protected static final String DEFAULT_KEYSPACE = "default_keyspace";
